[Add] 이름으로 병원검색 기능 개발 searchByHospitalName 기능의 repository 쿼리 작성, service 개발

+++ back/src/main/java/com/roller/doc/db/entity/HospitalPart.java
@@ -27,7 +27,7 @@ public class HospitalPart {
 	@Id
 	@Column(name="hospital_part_id", nullable = false)
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long hospital_part_id;
+	private int hospital_part_id;
 
 	@Column(name="hospital_part_name",columnDefinition = "INT")
 	private int hospital_part_name;
